Add kas_widgets::edit::Editor
