+++ crates/kas-widgets/src/edit/edit_field.rs
@@ -6,14 +6,14 @@
 //! The [`EditField`] and [`EditBox`] widgets, plus supporting items
 
 use super::*;
-use kas::event::components::{TextInput, TextInputAction};
-use kas::event::{CursorIcon, ElementState, FocusSource, PhysicalKey, Scroll};
-use kas::event::{Ime, ImePurpose, ImeSurroundingText};
+use kas::event::Ime;
+use kas::event::components::TextInputAction;
+use kas::event::{CursorIcon, ElementState, FocusSource, PhysicalKey};
 use kas::geom::Vec2;
 use kas::messages::{ReplaceSelectedText, SetValueText};
 use kas::prelude::*;
-use kas::text::{CursorRange, Effect, EffectFlags, NotReady, SelectionHelper};
-use kas::theme::{Text, TextClass};
+use kas::text::{CursorRange, Effect, EffectFlags, NotReady};
+use kas::theme::TextClass;
 use kas::util::UndoStack;
 use std::fmt::{Debug, Display};
 use std::ops::Range;
@@ -67,21 +67,15 @@ mod EditField {
     ///
     /// This is a [`Viewport`] widget.
     #[autoimpl(Debug where G: trait)]
+    #[autoimpl(Deref, DerefMut using self.editor)]
     #[widget]
     pub struct EditField<G: EditGuard = DefaultGuard<()>> {
         core: widget_core!(),
-        editable: bool,
         width: (f32, f32),
         lines: (f32, f32),
-        text: Text<String>,
-        selection: SelectionHelper,
-        edit_x_coord: Option<f32>,
+        editor: Editor,
         last_edit: Option<EditOp>,
         undo_stack: UndoStack<(String, CursorRange)>,
-        has_key_focus: bool,
-        current: CurrentAction,
-        error_state: bool,
-        input_handler: TextInput,
         /// The associated [`EditGuard`] implementation
         pub guard: G,
     }
@@ -121,6 +115,7 @@ mod EditField {
         }
 
         fn set_rect(&mut self, cx: &mut SizeCx, rect: Rect, mut hints: AlignHints) {
+            self.editor.pos = rect.pos;
             hints.vert = Some(if self.multi_line() {
                 Align::Default
             } else {
@@ -206,6 +201,7 @@ mod EditField {
         }
 
         fn configure(&mut self, cx: &mut ConfigCx) {
+            self.editor.id = self.id();
             self.text.configure(&mut cx.size_cx());
             G::configure(self, cx);
         }
@@ -414,7 +410,7 @@ mod EditField {
                     self.request_key_focus(cx, FocusSource::Pointer);
                     Used
                 }
-                event => match self.input_handler.handle(cx, self.id(), event) {
+                event => match self.editor.input_handler.handle(cx, self.core.id(), event) {
                     TextInputAction::Used => Used,
                     TextInputAction::Unused => Unused,
                     TextInputAction::PressStart {
@@ -432,7 +428,9 @@ mod EditField {
                         self.set_cursor_from_coord(cx, coord);
                         self.selection.set_anchor(clear);
                         if repeats > 1 {
-                            self.selection.expand(&self.text, repeats >= 3);
+                            self.editor
+                                .selection
+                                .expand(&self.editor.text, repeats >= 3);
                         }
 
                         self.request_key_focus(cx, FocusSource::Pointer);
@@ -442,7 +440,9 @@ mod EditField {
                         if self.current == CurrentAction::Selection {
                             self.set_cursor_from_coord(cx, coord);
                             if repeats > 1 {
-                                self.selection.expand(&self.text, repeats >= 3);
+                                self.editor
+                                    .selection
+                                    .expand(&self.editor.text, repeats >= 3);
                             }
                         }
 
@@ -505,18 +505,11 @@ mod EditField {
         pub fn new(guard: G) -> EditField<G> {
             EditField {
                 core: Default::default(),
-                editable: true,
                 width: (8.0, 16.0),
                 lines: (1.0, 1.0),
-                text: Text::new(String::new(), TextClass::Editor, false),
-                selection: Default::default(),
-                edit_x_coord: None,
+                editor: Editor::new(),
                 last_edit: Some(EditOp::Initial),
                 undo_stack: UndoStack::new(),
-                has_key_focus: false,
-                current: CurrentAction::None,
-                error_state: false,
-                input_handler: Default::default(),
                 guard,
             }
         }
@@ -588,7 +581,8 @@ mod EditField {
                 return false;
             }
 
-            self.selection.set_max_len(self.text.str_len());
+            let len = self.text.str_len();
+            self.selection.set_max_len(len);
             self.edit_x_coord = None;
             cx.redraw(&self);
             self.set_error_state(cx, false);
@@ -631,137 +625,6 @@ mod EditField {
             self.edit_x_coord = None;
             self.selection = range.into().into();
         }
-
-        /// Enable IME if not already enabled
-        fn enable_ime(&mut self, cx: &mut EventCx) {
-            if self.current.is_none() {
-                let hint = Default::default();
-                let purpose = ImePurpose::Normal;
-                let surrounding_text = self.ime_surrounding_text();
-                cx.request_ime_focus(self.id(), hint, purpose, surrounding_text);
-            }
-        }
-
-        /// Cancel on-going selection and IME actions
-        ///
-        /// This should be called if e.g. key-input interrupts the current
-        /// action.
-        fn cancel_selection_and_ime(&mut self, cx: &mut EventState) {
-            if self.current == CurrentAction::Selection {
-                self.input_handler.stop_selecting();
-                self.current = CurrentAction::None;
-            } else if self.current.is_ime_enabled() {
-                self.clear_ime();
-                cx.cancel_ime_focus(self.id_ref());
-            }
-        }
-
-        /// Clean up IME state
-        ///
-        /// One should also call [`EventCx::cancel_ime_focus`] unless this is
-        /// implied.
-        fn clear_ime(&mut self) {
-            if self.current.is_ime_enabled() {
-                let action = std::mem::replace(&mut self.current, CurrentAction::None);
-                if let CurrentAction::ImePreedit { edit_range } = action {
-                    self.selection.set_cursor(edit_range.start.cast());
-                    self.text.replace_range(edit_range.cast(), "");
-                }
-            }
-        }
-
-        fn ime_surrounding_text(&self) -> Option<ImeSurroundingText> {
-            const MAX_TEXT_BYTES: usize = ImeSurroundingText::MAX_TEXT_BYTES;
-
-            let sel_range = self.selection.range();
-            let edit_range = match self.current.clone() {
-                CurrentAction::ImePreedit { edit_range } => Some(edit_range.cast()),
-                _ => None,
-            };
-            let mut range = edit_range.clone().unwrap_or(sel_range);
-            let initial_range = range.clone();
-            let edit_len = edit_range.clone().map(|r| r.len()).unwrap_or(0);
-
-            if let Ok(Some((_, line_range))) = self.text.find_line(range.start) {
-                range.start = line_range.start;
-            }
-            if let Ok(Some((_, line_range))) = self.text.find_line(range.end) {
-                range.end = line_range.end;
-            }
-
-            if range.len() - edit_len > MAX_TEXT_BYTES {
-                range.end = range.end.min(initial_range.end + MAX_TEXT_BYTES / 2);
-                while !self.as_str().is_char_boundary(range.end) {
-                    range.end -= 1;
-                }
-
-                if range.len() - edit_len > MAX_TEXT_BYTES {
-                    range.start = range.start.max(initial_range.start - MAX_TEXT_BYTES / 2);
-                    while !self.as_str().is_char_boundary(range.start) {
-                        range.start += 1;
-                    }
-                }
-            }
-
-            let start = range.start;
-            let mut text = String::with_capacity(range.len() - edit_len);
-            if let Some(er) = edit_range {
-                text.push_str(&self.as_str()[range.start..er.start]);
-                text.push_str(&self.as_str()[er.end..range.end]);
-            } else {
-                text = self.as_str()[range].to_string();
-            }
-
-            let cursor = self.selection.edit_index().saturating_sub(start);
-            // Terminology difference: our sel_index is called 'anchor'
-            // SelectionHelper::anchor is not the same thing.
-            let sel_index = self.selection.sel_index().saturating_sub(start);
-            ImeSurroundingText::new(text, cursor, sel_index)
-                .inspect_err(|err| {
-                    // TODO: use Display for err not Debug
-                    log::warn!("EditField::ime_surrounding_text failed: {err:?}")
-                })
-                .ok()
-        }
-
-        // Call only if self.ime_focus
-        fn set_ime_cursor_area(&self, cx: &mut EventState) {
-            if let Ok(text) = self.text.display() {
-                let range = match self.current.clone() {
-                    CurrentAction::ImeStart => self.selection.range(),
-                    CurrentAction::ImePreedit { edit_range } => edit_range.cast(),
-                    _ => return,
-                };
-
-                let (m1, m2);
-                if range.is_empty() {
-                    let mut iter = text.text_glyph_pos(range.start);
-                    m1 = iter.next();
-                    m2 = iter.next();
-                } else {
-                    m1 = text.text_glyph_pos(range.start).next_back();
-                    m2 = text.text_glyph_pos(range.end).next();
-                }
-
-                let rect = if let Some((c1, c2)) = m1.zip(m2) {
-                    let left = c1.pos.0.min(c2.pos.0);
-                    let right = c1.pos.0.max(c2.pos.0);
-                    let top = (c1.pos.1 - c1.ascent).min(c2.pos.1 - c2.ascent);
-                    let bottom = (c1.pos.1 - c1.descent).max(c2.pos.1 - c2.ascent);
-                    let p1 = Vec2(left, top).cast_floor();
-                    let p2 = Vec2(right, bottom).cast_ceil();
-                    Rect::from_coords(p1, p2)
-                } else if let Some(c) = m1.or(m2) {
-                    let p1 = Vec2(c.pos.0, c.pos.1 - c.ascent).cast_floor();
-                    let p2 = Vec2(c.pos.0, c.pos.1 - c.descent).cast_ceil();
-                    Rect::from_coords(p1, p2)
-                } else {
-                    return;
-                };
-
-                cx.set_ime_cursor_area(self.id_ref(), rect + Offset::conv(self.rect().pos));
-            }
-        }
     }
 }
 
@@ -769,12 +632,8 @@ impl<A: 'static> EditField<DefaultGuard<A>> {
     /// Construct an `EditField` with the given inital `text` (no event handling)
     #[inline]
     pub fn text<S: ToString>(text: S) -> Self {
-        let text = text.to_string();
-        let len = text.len();
         EditField {
-            editable: true,
-            text: Text::new(text, TextClass::Editor, false),
-            selection: SelectionHelper::from(len),
+            editor: Editor::from(text),
             ..Default::default()
         }
     }
@@ -1320,9 +1179,9 @@ impl<G: EditGuard> EditField<G> {
             }
             Action::UndoRedo(redo) => {
                 if let Some((text, cursor)) = self.undo_stack.undo_or_redo(redo) {
-                    if self.text.set_str(text) {
-                        self.edit_x_coord = None;
-                        self.error_state = false;
+                    if self.editor.text.set_str(text) {
+                        self.editor.edit_x_coord = None;
+                        self.editor.error_state = false;
                     }
                     self.selection = (*cursor).into();
                     EditAction::Edit
@@ -1345,45 +1204,5 @@ impl<G: EditGuard> EditField<G> {
                 Used
             }
         })
-    }
-
-    // Set cursor position. It is assumed that the text has not changed.
-    fn set_cursor_from_coord(&mut self, cx: &mut EventCx, coord: Coord) {
-        let rel_pos = (coord - self.rect().pos).cast();
-        if let Ok(index) = self.text.text_index_nearest(rel_pos) {
-            if index != self.selection.edit_index() {
-                self.selection.set_edit_index(index);
-                self.set_view_offset_from_cursor(cx);
-                self.edit_x_coord = None;
-                cx.redraw();
-            }
-        }
-    }
-
-    fn set_primary(&self, cx: &mut EventCx) {
-        if self.has_key_focus && !self.selection.is_empty() && cx.has_primary() {
-            let range = self.selection.range();
-            cx.set_primary(String::from(&self.text.as_str()[range]));
-        }
-    }
-
-    /// Update view_offset after the cursor index changes
-    ///
-    /// It is assumed that the text has not changed.
-    ///
-    /// A redraw is assumed since the cursor moved.
-    fn set_view_offset_from_cursor(&mut self, cx: &mut EventCx) {
-        let cursor = self.selection.edit_index();
-        if let Some(marker) = self
-            .text
-            .text_glyph_pos(cursor)
-            .ok()
-            .and_then(|mut m| m.next_back())
-        {
-            let y0 = (marker.pos.1 - marker.ascent).cast_floor();
-            let pos = self.rect().pos + Offset(marker.pos.0.cast_nearest(), y0);
-            let size = Size(0, i32::conv_ceil(marker.pos.1 - marker.descent) - y0);
-            cx.set_scroll(Scroll::Rect(Rect { pos, size }));
-        }
     }
 }
